added Constants.py and date + count based file writing to avoid overwriting
codes.py was modified to create Constants.py. Constants.py contains
constants, error codes, strings etc that is to be used throughout the
appliation. It is a central key manager and localization unit.

Also added date and count of all images previously posted by the user
to the path of the file. Now users can upload multiple files without
overwriting.

Edit: had to change server IP, because of change of local IP.