modified code to work with google cloud platform. ie. google app engine. Added some test pythons which are not useful to the project but to test various functions of python + google cloud storage.
For much time I was facing the problem of image file writing. Google App
Engine does not allow image to be written to the same location as the
program for security reasons. So Images have to be written to cloud sto-
rage buckets. There were a number of tutorials and documentations on how
to do that with python and google cloud storage, but they were lacking in
some details which confused me. But in the end I was able to install the
nececcary packages required for google cloud storage.

The static directory was created using django commands to aid google app
engine

Database was migrated to use google's sql database. it uses cloud_sql_proxy
to be able to connect to the sql from the local computer server.