Added a privacy column for User table, search_username function for views and some constants.
The privacy column is used for handling privacy actions. All other
functions have to be modified to include privacy : for example,
fetch_post can check privacy as a security measure, or fetch_feeds
can check and return only those feeds followed by the user etc ...

Also changed the constants and how they are shared (im not sure if
I commited this change before? :/ , so here it is). The code has
been changed to use Constants and ECodes as part of a dict object.
Which is converted to JSON and sent to the user device when it loads
the application. Thus allowing handline of ecodes from the server.
The codes are obtained in program by use of getCode(name) which
returns the code corresponding to the `name`.